frontend/requirements.txt
gradio==4.44.0
huggingface_hub==0.26.2
requests
Pillow
